requirements.txt
streamlit==1.38.0
groq==0.9.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
